requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
Pillow>=10.0.0
google-genai>=1.47.0
python-dotenv>=1.0.0
Markdown>=3.9
pytest>=7.4.0
pytest-cov>=4.1.0

